Adicionado codigo do motor de passo
- Codigo do motor de passo esta em src/eletronica/motor_passo.ino
  e deve ser rodado com o rosserial

- script rebuild_ws.sh agora deleta devel/ e build/ sem dar erro quando as
  pastas nao existem

+++ autobotz_ws/rebuild_ws.sh
@@ -1,6 +1,16 @@
-rm src/CMakeLists.txt
+if [ -d "devel" ]; then
+	rm -r devel/
+fi
+
+if [ -d "build" ]; then
+	rm -r build/
+fi
+
+if [ -e "src/CMakeLists.txt" ]; then
+	rm src/CMakeLists.txt
+fi
 
 source /opt/ros/indigo/setup.bash
 
 catkin_make
-
+catkin_make
